# Refactor render events