--- a/osu.Framework/Graphics/Rendering/Deferred/Events/ClearEvent.cs
+++ b/osu.Framework/Graphics/Rendering/Deferred/Events/ClearEvent.cs
@@ -6,6 +6,6 @@ namespace osu.Framework.Graphics.Rendering.Deferred.Events
     internal readonly record struct ClearEvent(ClearInfo Info)
     {
         public static RenderEvent Create(ClearInfo info)
-            => RenderEvent.Init(new ClearEvent(info));
+            => RenderEvent.Create(new ClearEvent(info));
     }
 }