--- a/osu.Framework/Graphics/Rendering/Deferred/Events/SetBlendEvent.cs
+++ b/osu.Framework/Graphics/Rendering/Deferred/Events/SetBlendEvent.cs
@@ -6,6 +6,6 @@ namespace osu.Framework.Graphics.Rendering.Deferred.Events
     internal readonly record struct SetBlendEvent(BlendingParameters Parameters)
     {
         public static RenderEvent Create(BlendingParameters parameters)
-            => RenderEvent.Init(new SetBlendEvent(parameters));
+            => RenderEvent.Create(new SetBlendEvent(parameters));
     }
 }