--- a/osu.Framework/Graphics/Rendering/Deferred/Events/SetBlendMaskEvent.cs
+++ b/osu.Framework/Graphics/Rendering/Deferred/Events/SetBlendMaskEvent.cs
@@ -6,6 +6,6 @@ namespace osu.Framework.Graphics.Rendering.Deferred.Events
     internal readonly record struct SetBlendMaskEvent(BlendingMask Mask)
     {
         public static RenderEvent Create(BlendingMask mask)
-            => RenderEvent.Init(new SetBlendMaskEvent(mask));
+            => RenderEvent.Create(new SetBlendMaskEvent(mask));
     }
 }