--- a/osu.Framework/Graphics/Rendering/Deferred/Events/SetFrameBufferEvent.cs
+++ b/osu.Framework/Graphics/Rendering/Deferred/Events/SetFrameBufferEvent.cs
@@ -8,6 +8,6 @@ namespace osu.Framework.Graphics.Rendering.Deferred.Events
     internal readonly record struct SetFrameBufferEvent(ResourceReference FrameBuffer)
     {
         public static RenderEvent Create(DeferredRenderer renderer, IFrameBuffer? frameBuffer)
-            => RenderEvent.Init(new SetFrameBufferEvent(renderer.Context.Reference(frameBuffer)));
+            => RenderEvent.Create(new SetFrameBufferEvent(renderer.Context.Reference(frameBuffer)));
     }
 }