--- a/osu.Framework/Graphics/Rendering/Deferred/Events/SetScissorEvent.cs
+++ b/osu.Framework/Graphics/Rendering/Deferred/Events/SetScissorEvent.cs
@@ -8,6 +8,6 @@ namespace osu.Framework.Graphics.Rendering.Deferred.Events
     internal readonly record struct SetScissorEvent(RectangleI Scissor)
     {
         public static RenderEvent Create(RectangleI scissor)
-            => RenderEvent.Init(new SetScissorEvent(scissor));
+            => RenderEvent.Create(new SetScissorEvent(scissor));
     }
 }