--- a/osu.Framework/Graphics/Rendering/Deferred/Events/SetShaderEvent.cs
+++ b/osu.Framework/Graphics/Rendering/Deferred/Events/SetShaderEvent.cs
@@ -9,6 +9,6 @@ namespace osu.Framework.Graphics.Rendering.Deferred.Events
     internal readonly record struct SetShaderEvent(ResourceReference Shader)
     {
         public static RenderEvent Create(DeferredRenderer renderer, IShader shader)
-            => RenderEvent.Init(new SetShaderEvent(renderer.Context.Reference(shader)));
+            => RenderEvent.Create(new SetShaderEvent(renderer.Context.Reference(shader)));
     }
 }