… osu.Framework/Graphics/Rendering/Deferred/Events/SetShaderStorageBufferObjectDataEvent.cs
… osu.Framework/Graphics/Rendering/Deferred/Events/SetStencilInfoEvent.cs
… osu.Framework/Graphics/Rendering/Deferred/Events/SetTextureEvent.cs
… osu.Framework/Graphics/Rendering/Deferred/Events/SetUniformBufferDataEvent.cs
… osu.Framework/Graphics/Rendering/Deferred/Events/SetUniformBufferDataRangeEvent.cs
… osu.Framework/Graphics/Rendering/Deferred/Events/SetUniformBufferEvent.cs
… osu.Framework/Graphics/Rendering/Deferred/Events/SetViewportEvent.cs