See [this sharplab](https://sharplab.io/#v2:C4LghgzgtgPgAgBgARwIwDoBKBXAdsASygFN0BhAeygAcCAbYgJwGUmA3AgY2IgG4BYAFCIUGHPiKkAkviYVqrRh258hQgNrNgjbJ2AAZMAE8K2YAApDJswGkCuACboAogA9qdLgWABKALpCSEHBQXAAzEgQ2rrASAAqPMBaOnqBIcEA3mnp6eoAYgTEdA4A8gBmZRDEFgj+2TnB4UgARkbAxEgAggKC9Q1I+YXF5ZXV5rUBgv0hTfaxAEI9fQ2DRaUVVTV1U9OhEXQUuADmSGRLO7urwxtjABzbu3stbR0AIuePAwVrI5vm95NPrN8EhnB9Hld1qMLADljkmgdjkg8uDLt9rtDzKgAGwPR5NVrtJAAcVR00hvzGOLxu2BsQAEmT+hSbhZqYD8ftDicpEyGk0ElFkjFzISOmAfHD0lkLp8ukgALxIACsPU+AF81LL+k00Nj4olhXokDJvK8CIxiHpRS8kBLFQA+JC4YgAdwNQuiqW1DRlcuCnUVKqlIXVfPhET1HqSXtipuAAHVvAALACquAgYDKxGYAGsCNR4zaiRKQ5kyzl05ns+g8wWi6ZYoKYylYhAfGr/ZF0IGlaqK+k4AB2SKd3aa3o+nKafPUfQUThgOgQeMc2mRnHRo1x3DeJPAZN1wu74CdYDaAjNMzEYvinyO51urexgd+ru94NT9JhqXqoA==) for a showcase. Of note:
1. Explicitly implemented ctor will zero-initialise each individual field. `RenderEvent` is a pretty big struct (by number of fields, not by bytes), so we don't want this.
2. There appears to be no difference with `[SkipLocalsInit]` attached to the initialiser implementation used in this PR.
3. `Unsafe.SkipInit()` (the previous implementation) is slower due to code size. The benchmarks below show differences surfacing likely as a result of lack of inlining. This is why I consider this a minor performance-related PR.

Before:

|                           Method |     Mean |    Error |   StdDev |
|--------------------------------- |---------:|---------:|---------:|
|              AddPrimitiveToBatch | 67.99 us | 0.020 us | 0.017 us |
|                            Clear | 87.04 us | 0.032 us | 0.027 us |
|                   DrawNodeAction | 27.20 us | 0.007 us | 0.007 us |
|                            Flush | 27.24 us | 0.036 us | 0.030 us |
|                ResizeFrameBuffer | 87.10 us | 0.063 us | 0.059 us |
|                         SetBlend | 87.07 us | 0.027 us | 0.022 us |
|                     SetBlendMask | 24.50 us | 0.018 us | 0.017 us |
|                     SetDepthInfo | 68.02 us | 0.050 us | 0.047 us |
|                   SetFrameBuffer | 24.49 us | 0.024 us | 0.020 us |
|                       SetScissor | 67.99 us | 0.026 us | 0.022 us |
|                        SetShader | 24.51 us | 0.014 us | 0.013 us |
|                  SetScissorState | 24.51 us | 0.024 us | 0.022 us |
| SetShaderStorageBufferObjectData | 95.20 us | 0.025 us | 0.022 us |
|                   SetStencilInfo | 87.05 us | 0.039 us | 0.035 us |
|                       SetTexture | 27.22 us | 0.016 us | 0.014 us |
|             SetUniformBufferData | 87.07 us | 0.081 us | 0.076 us |
|        SetUniformBufferDataRange | 87.06 us | 0.041 us | 0.038 us |
|                 SetUniformBuffer | 27.22 us | 0.027 us | 0.024 us |
|                      SetViewport | 71.11 us | 0.225 us | 0.211 us |

After:

|                           Method |     Mean |    Error |   StdDev |
|--------------------------------- |---------:|---------:|---------:|
|              AddPrimitiveToBatch | 21.77 us | 0.014 us | 0.013 us |
|                            Clear | 21.78 us | 0.019 us | 0.016 us |
|                   DrawNodeAction | 27.22 us | 0.012 us | 0.010 us |
|                            Flush | 27.30 us | 0.043 us | 0.039 us |
|                ResizeFrameBuffer | 21.85 us | 0.035 us | 0.033 us |
|                         SetBlend | 21.84 us | 0.041 us | 0.037 us |
|                     SetBlendMask | 24.55 us | 0.037 us | 0.035 us |
|                     SetDepthInfo | 21.79 us | 0.026 us | 0.023 us |
|                   SetFrameBuffer | 24.60 us | 0.059 us | 0.055 us |
|                       SetScissor | 21.86 us | 0.040 us | 0.037 us |
|                        SetShader | 24.59 us | 0.047 us | 0.040 us |
|                  SetScissorState | 24.56 us | 0.036 us | 0.032 us |
| SetShaderStorageBufferObjectData | 21.78 us | 0.015 us | 0.013 us |
|                   SetStencilInfo | 21.79 us | 0.024 us | 0.022 us |
|                       SetTexture | 27.22 us | 0.026 us | 0.023 us |
|             SetUniformBufferData | 21.80 us | 0.041 us | 0.036 us |
|        SetUniformBufferDataRange | 21.80 us | 0.038 us | 0.034 us |
|                 SetUniformBuffer | 27.24 us | 0.025 us | 0.023 us |
|                      SetViewport | 21.78 us | 0.016 us | 0.014 us |